Factor out MessagesFilters into its own module
