+++ frontend/Messages.tsx
@@ -3,9 +3,9 @@ import { For, Show, JSX, createMemo } from "solid-js";
 import { Table, Thead, Tbody, Tr, Th, Td, Tooltip, Text } from "@hope-ui/solid";
 import { useSearchParams } from "solid-app-router";
 
-import { getMessages, instances } from "./store";
+import { getMessages } from "./store";
 import type { Message, UpdateMessage } from "../types/message";
-import Filter from "./Filter";
+import Filters from "./MessagesFilters";
 
 const COLUMN_WIDTHS = {
   instanceId: "4%",
@@ -94,69 +94,6 @@ const MessageComponent: Component<{ message: Message }> = (props) => {
         <UpdateMessageComponent message={props.message as UpdateMessage} />
       </Show>
     </Tr>
-  );
-};
-
-const instanceIdEntries = createMemo(() => {
-  const resolvedInstances = instances();
-  const all_entry = {
-    value: "*",
-    text: "All instance ids",
-  };
-  if (resolvedInstances == null) {
-    return [all_entry];
-  }
-  return [
-    all_entry,
-    ...resolvedInstances.map((instance) => ({
-      value: instance.id,
-      text: instance.id,
-    })),
-  ];
-});
-
-type ParamsType = ReturnType<typeof useSearchParams>[0];
-type SetParamsType = ReturnType<typeof useSearchParams>[1];
-
-const Filters: Component<{
-  searchParams: ParamsType;
-  setSearchParams: SetParamsType;
-}> = (props) => {
-  return (
-    <>
-      <Filter
-        label="instanceId"
-        entries={instanceIdEntries()}
-        value={props.searchParams.instanceId || "*"}
-        onChange={(value) => {
-          if (value === "*") {
-            props.setSearchParams({
-              ...props.searchParams,
-              instanceId: undefined,
-            });
-          } else {
-            props.setSearchParams({ ...props.searchParams, instanceId: value });
-          }
-        }}
-      />
-      <Filter
-        label="type"
-        entries={[
-          { value: "*", text: "All types" },
-          { value: "sent", text: "Sent" },
-          { value: "received", text: "Received" },
-          { value: "clear", text: "Clear" },
-        ]}
-        value={props.searchParams.type || "*"}
-        onChange={(value) => {
-          if (value === "*") {
-            props.setSearchParams({ ...props.searchParams, type: undefined });
-          } else {
-            props.setSearchParams({ ...props.searchParams, type: value });
-          }
-        }}
-      />
-    </>
   );
 };
